requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
filelock==3.0.12
huggingface-hub==0.0.12
idna==3.2
joblib==1.0.1
numpy==1.21.1
packaging==21.0
pyparsing==2.4.7
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.0
tqdm==4.61.2
transformers==4.8.2
typing-extensions==3.10.0.0
urllib3==1.26.6
